feat: add deposit and withdraw operation, and list accounts pagination, add handling exceptions for deposit and withdraw operaitons, add new handling exception for deleting bank that contains accounts

+++ src/main/java/com/sanad/firstspringbootproject/service/BankService.java
@@ -1,10 +1,12 @@
-package com.sanad.firstspringbootproject.service.bank;
+package com.sanad.firstspringbootproject.service;
 
 import com.sanad.firstspringbootproject.dto.bank.BankResponse;
-import com.sanad.firstspringbootproject.exception.bank.BankNotFoundException;
-import com.sanad.firstspringbootproject.exception.bank.DuplicateBankException;
+import com.sanad.firstspringbootproject.exception.BankHasAccountsException;
+import com.sanad.firstspringbootproject.exception.BankNotFoundException;
+import com.sanad.firstspringbootproject.exception.DuplicateBankException;
 import com.sanad.firstspringbootproject.mapper.BankMapper;
 import com.sanad.firstspringbootproject.model.Bank;
+import com.sanad.firstspringbootproject.repository.SpringAccountRepository;
 import com.sanad.firstspringbootproject.repository.SpringDataBankRepository;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
@@ -17,12 +19,14 @@ import java.util.Locale;
 @Transactional(readOnly = true)
 public class BankService {
 
+    private final SpringAccountRepository accountRepository;
     private final SpringDataBankRepository bankRepository;
     private final BankMapper bankMapper;
 
-    public BankService(SpringDataBankRepository bankRepository, BankMapper bankMapper) {
+    public BankService(SpringDataBankRepository bankRepository, SpringAccountRepository accountRepository ,BankMapper bankMapper) {
         this.bankRepository = bankRepository;
         this.bankMapper = bankMapper;
+        this.accountRepository = accountRepository;
     }
 
     public List<BankResponse> findAll() {
@@ -76,7 +80,16 @@ public class BankService {
     @Transactional
     public void delete(Long id) {
         Bank bank = findEntityById(id);
-        bankRepository.delete(bank);
+
+        if (accountRepository.existsByBankId(id)){
+            throw new BankHasAccountsException(id);
+        }
+        try {
+            bankRepository.delete(bank);
+            bankRepository.flush();
+        } catch (DataIntegrityViolationException exception) {
+            throw new BankHasAccountsException(id);
+        }
     }
 
     private Bank findEntityById(Long id) {
